Refactor duration error handling: change InvalidDuration type from f64 to f32 for consistency in safe_duration_secs_to_millis function

+++ abop-core/src/utils/casting/error.rs
@@ -122,7 +122,7 @@ pub mod domain {
 
         /// Invalid animation duration value
         #[error("Invalid animation duration: {0}")]
-        InvalidDuration(f64),
+    InvalidDuration(f32),
 
         /// Invalid color value
         #[error("Invalid color value: {0}")]
